Move from source to header
